refactor:resolved the comments
