+++ src/pages/Home.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useReducer, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
-import { useBidContext } from '../Hooks/BidContext';
+import { useBidContext } from '../components/BidContext';
 import Button from '../components/Button';
 import CountdownTimer from '../components/CountDownTimer';
 import Footer from '../components/Footer';
